+++ backend/repository/task_test/setup_test.go
@@ -1,0 +1,24 @@
+package task_test
+
+import (
+    "go-react-app/model"
+    "go-react-app/repository"
+    "go-react-app/testutils"
+    "gorm.io/gorm"
+)
+
+var (
+    taskDB *gorm.DB
+    taskRepo repository.ITaskRepository
+    taskTestUser model.User
+    taskOtherUser model.User
+    nonExistentTaskID uint = 9999
+)
+
+func setupTaskTest() {
+    taskDB = testutils.SetupTestDB()
+    taskRepo = repository.NewTaskRepository(taskDB)
+    
+    taskTestUser = testutils.CreateTestUser(taskDB)
+    taskOtherUser = testutils.CreateOtherUser(taskDB)
+}
